+++ src/background/menu.ts
@@ -1,0 +1,34 @@
+const menuTemplateForMac: Electron.MenuItemConstructorOptions[] = [
+  {
+    label: "VOICEVOX",
+    submenu: [{ role: "quit" }],
+  },
+  {
+    label: "Edit",
+    submenu: [
+      { role: "cut" },
+      { role: "copy" },
+      { role: "paste" },
+      { role: "selectAll" },
+    ],
+  },
+];
+
+const menuTemplateView: Electron.MenuItemConstructorOptions = {
+  label: "View",
+  submenu: [
+    { role: "zoomIn" },
+    { role: "zoomOut" },
+    { role: "toggleDevTools" },
+  ],
+};
+
+/**
+ * electron用のメニューのテンプレートを生成する
+ */
+export function generateMenuTemplate({ isMac }: { isMac: boolean }) {
+  const menuTemplate = isMac
+    ? [...menuTemplateForMac, menuTemplateView]
+    : [menuTemplateView];
+  return menuTemplate;
+}
